Changed exit to remove from tree and parents in cache instead of the same moment
